+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Fix url used for fetching streams
+- Return `is_end_sequence` on stream fetch
 
 ## v0.20.0