+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Fix url used for fetching streams
 - Return `is_end_sequence` on stream fetch
+- Make `transform_tag` optional on `create bucket`
 
 ## v0.20.0