+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Fix url used for fetching streams
 - Return `is_end_sequence` on stream fetch
 - Make `transform_tag` optional on `create bucket`
+- Retry `put_emails` requests
 
 ## v0.20.0